Added marshal Json method to Place, Profile, Relationship and Tombstone

+++ collection.go
@@ -329,7 +329,7 @@ func (c Collection) MarshalJSON() ([]byte, error) {
 		notEmpty = writeItemCollectionProp(&b, "items", c.Items) || notEmpty
 	}
 	writeCommaIfNotEmpty(notEmpty)
-	notEmpty = writeIntProp(&b, "totalItems", int(c.TotalItems)) || notEmpty
+	notEmpty = writeIntProp(&b, "totalItems", int64(c.TotalItems)) || notEmpty
 
 	if notEmpty {
 		b.Write([]byte{'}'})
