+++ ordered_collection_page.go
@@ -307,7 +307,7 @@ func (c OrderedCollectionPage) MarshalJSON() ([]byte, error) {
 	}
 
 	writeCommaIfNotEmpty(notEmpty)
-	notEmpty = writeIntProp(&b, "totalItems", int(c.TotalItems)) || notEmpty
+	notEmpty = writeIntProp(&b, "totalItems", int64(c.TotalItems)) || notEmpty
 
 	if notEmpty {
 		b.Write([]byte{'}'})
